First Step to Converting Platonic Solid Gen to Jobs
Tetrahedron now uses as Simple, Hard-Coded (memo-ized if I am being generous) BuildGraph function, which builds a native graph; bypassing the ProceduralMeshGen pipeline and instead building the mesh as a graph directly.

This will allow me to fix issue #1 by working at a polygonal level.
However; to optimize subdivisions; I plan to avoid doing binary subdivisions, and instead manually iterate the subdivisions, and I may not be able to do this for pentagons.